Feature/8.5 jwt account

+++ ProjectISS/src/main/java/com/booking/ProjectISS/controller/users/GuestController.java
@@ -9,6 +9,7 @@ import com.booking.ProjectISS.dto.users.GuestDTO;
 import com.booking.ProjectISS.dto.users.OwnerDTO;
 import com.booking.ProjectISS.model.accomodations.Accommodation;
 import com.booking.ProjectISS.model.reviews.Review;
+import com.booking.ProjectISS.model.users.Administrator;
 import com.booking.ProjectISS.model.users.Guest;
 import com.booking.ProjectISS.model.users.Owner;
 import com.booking.ProjectISS.service.accommodation.IAccommodationService;
@@ -27,10 +28,11 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Date;
 import java.util.List;
+import java.util.stream.Collectors;
 
-@CrossOrigin(origins = "http://localhost:4200")
 @RestController
 @RequestMapping("/api/guests")
+@CrossOrigin(origins = "http://localhost:4200")
 public class GuestController {
 
     @Autowired
@@ -64,6 +66,12 @@ public class GuestController {
         return new ResponseEntity<GuestDTO>(HttpStatus.NOT_FOUND);
     }
 
+    @GetMapping(value = "/username/{username}", produces = MediaType.APPLICATION_JSON_VALUE)
+
+    public ResponseEntity<Guest> getGuestUsername(@PathVariable("username") String username){
+        return new ResponseEntity<Guest>(guestService.findUsername(username), HttpStatus.OK);
+    }
+
     //delete one, 3.4 for guest
 //    @GetMapping(value = "/{id}/favouriteAccommodations", produces = MediaType.APPLICATION_JSON_VALUE)
 //    public ResponseEntity<Collection<Accommodation>> getFavouriteAccommodation(@PathVariable("id") Long id){
@@ -85,6 +93,7 @@ public class GuestController {
         if(status != null){
             reservations = reservationService.filterReservations(reservations, status);
         }
+
 
         return new ResponseEntity<Collection<ReservationDTO>>(reservationService.getReservationsDTO(reservations),
         HttpStatus.OK);
@@ -108,13 +117,18 @@ public class GuestController {
     //deleteOne
     @DeleteMapping(value = "/{id}")
     public ResponseEntity<GuestDTO> deleteGuest(@PathVariable("id") Long id) {
+        List<Reservation> acceptedReservations = reservationService.getGuestReservations(id).stream()
+                .filter(reservation -> reservation.getStatus() == ReservationStatus.ACCEPTED)
+                .toList();
+
+        if(!acceptedReservations.isEmpty()){return  new ResponseEntity<>(HttpStatus.FORBIDDEN);}
         guestService.delete(id);
         return new ResponseEntity<GuestDTO>(HttpStatus.NO_CONTENT);
     }
 
-    @CrossOrigin(origins = "http://localhost:4200")
-    @DeleteMapping(value = "/request/{reqId}")
-    public ResponseEntity<ReservationDTO> deleteGuestReservation(@PathVariable("reqId") Long reqId){
+    @DeleteMapping(value = "/{id}/request/{reqId}")
+    public ResponseEntity<ReservationDTO> deleteGuestReservation(@PathVariable("id") Long id,
+                                                                 @PathVariable("reqId") Long reqId){
         if(reservationService.findOne(reqId).getStatus() == ReservationStatus.ACCEPTED){
             return new ResponseEntity<ReservationDTO>(HttpStatus.NOT_FOUND);
         }else{
@@ -176,6 +190,9 @@ public class GuestController {
         return new ResponseEntity<>("Guest cant report", HttpStatus.BAD_REQUEST);
     }
 
+
+
+
 //    @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 //    public ResponseEntity<GuestDTO> createGuest(@RequestBody Guest guest) throws Exception {
 //        GuestDTO guestDTO = guestService.create(guest);
@@ -212,11 +229,5 @@ public class GuestController {
         System.out.println("UDJE3");
         ReservationDTO reservationDTO = reservationService.create(reservation);
         return new ResponseEntity<>(reservationDTO, HttpStatus.CREATED);
-    }
-    @CrossOrigin(origins = "http://localhost:4200")
-    @GetMapping(value = "/request")
-    public ResponseEntity<Collection<ReservationDTO>> guestNotAcceptedReservation(){
-        Collection<ReservationDTO> reservations = reservationService.findAllNotAcceptedDTO();
-        return new ResponseEntity<Collection<ReservationDTO>>(reservations, HttpStatus.OK);
     }
 }
